# api-spec-v3.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses/{id}/listings':
    get:
      tags:
        - 'U22: List sale'
        - 'Business Accounts'
        - 'Listings / sales'
      summary: "Retrieve a business's listings"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Listings retrieved successfully. This could be an empty array.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    post:
      tags:
        - 'U22: List sale'
        - 'Business Accounts'
        - 'Listings / sales'
      summary: "Add a new listing for some inventory item of a business"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Listing created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  listingId:
                    type: integer
                    description: The ID of the listing that has just been created
                    example: 57
        '400':
          description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
        '403':
          description: 'Forbidden: Returned when a user tries to add listing for a business they do not administer AND the user is not a global application admin'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Listing'
...
  '/businesses/listings':
    post:
      tags:
        - 'Listing / sales'
        - 'U29 : Browse/search sale listings'
      summary: Search across all business listings with (optional) given parameters.
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: query
          name: count
          schema:
            type: number
          description: 'How many results per page to return'
          example: '10'
        - in: query
          name: offset
          schema:
            type: number
          description: 'How many pages to skip before returning the results'
          example: '2'
        - in: query
          name: sortDirection
          schema:
            type: string
          description: 'Which direction to sort by: asc, desc. Invalid directions default to asc'
          example: 'desc'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessListingSearchRequest'
      responses:
        '200':
          description: Retrieves a (potentially empty) paginated list of listings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
...
    Listing:
      type: object
      description: A listing for the sale of some product from a business's inventory
      required:
        - inventoryItemId
        - quantity
        - price
      properties:
        id:
          type: integer
          example: 57
          readOnly: true
          description: A unique (within the business) identifer for this listing - autogenerated by the system
        inventoryItem:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/InventoryItem'
        inventoryItemId:
          type: integer
          example: 101
          writeOnly: true
        quantity:
          type: integer
          example: 3
          description: Must be greater than zero, and less than or equal to the quantity of the correpsonding inventory entry
        price:
          type: number
          format: double
          description: Price of the listing
          example: 17.99
        moreInfo:
          type: string
          example: 'Seller may be willing to consider near offers'
          description: An optional field for further information about the listing
        created:
          type: string
          format: date-time
          example: '2021-07-14T11:44:00Z'
          readOnly: true
          description: Date and time when the listing was created
        closes:
          type: string
          format: date-time
          example: '2021-07-21T23:59:00Z'
          description: Date and time when the listing closes (defaults to expiry date of related inventory item)
        likes:
          type: number
          format: integer
          example: 10
          description: The number of likes the listing has received (defaults to 0)
    InventoryItem:
      type: object
      description: Entity representing an instance of a stock of some product
      required:
        - productId
        - quantity
        - expires
      properties:
        id:
          type: integer
          example: 101
          readOnly: true
          description: A unique (within the inventory) identifer for this item - autogenerated by the system
        product:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/Product'
        productId:
          writeOnly: true
          type: string
          example: "WATT-420-BEANS"
          description: Product Code of corresponding product in business's catalogue. Notice that the ID is given when writing, but the full object is returned when reading
        quantity:
          type: integer
          example: 4
          description: How many items of stock are present
        pricePerItem:
          type: number
          format: double
          description: Price to purchase a single item of the product
          example: 6.50
        totalPrice:
          type: number
          format: double
          description: Price for purchasing the entire stock of the product
          example: 21.99
        manufactured:
          type: string
          format: date
          description: The date on which the product was prepared / manufactured
        sellBy:
          type: string
          format: date
          description: The date by which the product ought to be sold
        bestBefore:
          type: string
          format: date
          description: 'The "best before" date of the product'
        expires:
          type: string
          format: date
          description: The expiry date of the product
    Image:
      type: object
      description: Entity representing an image stored on the server
      required:
        - filename
        - thumbnailFilename
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
          description: A unique identifier (system-wide) to represent this image
        filename:
          type: string
          example: "/media/images/23987192387509-123908794328.png"
          description: Relative URI to full size image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
        thumbnailFilename:
          type: string
          example: "/media/images/23987192387509-123908794328_thumbnail.png"
          description: Relative URI to thumbnail of image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
    Product:
      type: object
      description: Entity for catalogue items in a product catalogue
      required:
        - name
      properties:
        id:
          type: string
          example: 'WATT-420-BEANS'
          description: >
            Developers can choose whether or not to make this a required field (i.e user
            specified vs. auto-generated). In the case that the team automatically
            generates the IDs, they should still allow this field to be present when
            creating (POST) or modifying (PUT) a product, and then simply ignore its
            value (i.e do not return a 400 code if it is present but not needed). An
            ID only needs to be unique within its catalogue, not system-wide.
        name:
          type: string
          description: The full name of the product
          example: "Watties Baked Beans - 420g can"
        description:
          type: string
          description: A desription of the product
          example: "Baked Beans as they should be."
        manufacturer:
          type: string
          description: The manufacturer of the product
          example: "Heinz Wattie's Limited"
        recommendedRetailPrice:
          type: number
          format: double
          example: 2.20
          description: The recommended retail price of the product, given in whatever currency is in use at the business's location.
        created:
          type: string
          format: date-time
          readOnly: true
          description: The date and time when this product was created. This field is generated automatically by the server.
        images:
          type: array
          readOnly: true
          description: List of images of a product (implemented in U16).
          items:
            $ref: '#/components/schemas/Image'
...
    BusinessListingSearchRequest:
      type: object
      properties:
        businessQuery:
          type: string
          description: 'Search for all, or part of the name of a business'
          example: 'Dabshot'
        productQuery:
          type: string
          description: 'Search for all, or part of the name of a seller'
          example: 'Apples'
        addressQuery:
          type: string
          description: 'Search for all, or part of the sellers address'
          example: 'Christchurch'
        sortBy:
          type: string
          description: 'Sort by a specific category'
          example: 'price'
        businessTypes:
          type: array
          items:
            type: string
            enum:
              - Accommodation and Food Services
              - Retail Trade
              - Charitable organisation
              - Non-profit organisation
          description: "Search for one or more business type(s)"
        minPrice:
          type: number
          format: double
          description: 'Limit the results by setting a minimum price range'
          example: '5.00'
        maxPrice:
          type: number
          format: double
          description: 'Limit the results by setting a maximum price range'
          example: '25.00'
        minClosingDate:
          type: string
          format: date
          description: 'Limit the results by setting a minimum closing date'
          example: '2021-09-09 12:00:00'
        maxClosingDate:
          type: string
          format: date
          description: 'Limit the results by setting a maximum closing date'
          example: '2021-09-12 12:00:00'
...